UI: Move Redirect under components
